requirements.txt
apex==0.9.10dev
diffusers==0.29.0
einops==0.8.0
kornia==0.6.5
nltk==3.8.1
numpy==1.23.0
Pillow==11.0.0
PyYAML==6.0.2
scikit_learn==1.4.1.post1
scipy==1.14.1
setuptools==60.2.0
termcolor==2.5.0
timm==0.3.2
torch==1.11.0+cu113
torchmetrics==1.4.0.post0
torchvision==0.12.0+cu113
tqdm==4.66.5
yacs==0.1.8
